// src/platform/packages/shared/kbn_esql_language/src/types.ts
export type ESQLAst = ESQLAstCommand[];

export type ESQLAstCommand =
  | ESQLCommand
  | ESQLAstJoinCommand
  | ESQLAstChangePointCommand
  | ESQLAstRerankCommand
  | ESQLAstCompletionCommand
  | ESQLAstFuseCommand
  | ESQLAstForkCommand;

export type ESQLAstAllCommands = ESQLAstCommand | ESQLAstHeaderCommand;

export type ESQLAstNode = ESQLAstCommand | ESQLAstHeaderCommand | ESQLAstExpression | ESQLAstItem;

/**
 * Represents an *expression* in the AST.
 */
export type ESQLAstExpression = ESQLSingleAstItem;

export type ESQLSingleAstItem =
  | ESQLAstQueryExpression
  | ESQLFunction
  | ESQLCommandOption
  | ESQLSource
  | ESQLParens
  | ESQLColumn
  | ESQLDatePeriodLiteral
  | ESQLTimeDurationLiteral
  | ESQLList
  | ESQLLiteral
  | ESQLIdentifier
  | ESQLInlineCast
  | ESQLOrderExpression
  | ESQLUnknownItem
  | ESQLMap
  | ESQLMapEntry
  | PromQLAstQueryExpression;
// ...
export type ESQLAstItem = ESQLSingleAstItem | ESQLAstItem[];
// ...
export interface ESQLLocation {
  min: number;
  max: number;
}

export interface ESQLAstBaseItem<Name = string> {
  name: Name;
  text: string;
  location: ESQLLocation;
  incomplete: boolean;
  formatting?: ESQLAstNodeFormatting;
}

/**
 * Contains optional formatting information used by the pretty printer.
 */
export interface ESQLAstNodeFormatting {
  top?: ESQLAstComment[];
  left?: ESQLAstCommentMultiLine[];
  right?: ESQLAstCommentMultiLine[];
  rightSingleLine?: ESQLAstCommentSingleLine;
  bottom?: ESQLAstComment[];
}

export interface ESQLCommand<Name = string> extends ESQLAstBaseItem<Name> {
  type: 'command';

  /**
   * The subtype of the command. For example, the `JOIN` command can be: (1)
   * LOOKUP JOIN, (2) LEFT JOIN, (3) RIGHT JOIN.
   */
  commandType?: string;

  args: ESQLAstItem[];
}

export interface ESQLAstJoinCommand extends ESQLCommand<'join'> {
  commandType: 'lookup' | 'left' | 'right';
}

export interface ESQLAstChangePointCommand extends ESQLCommand<'change_point'> {
  value: ESQLColumn;
  key?: ESQLColumn;
  target?: {
    type: ESQLColumn;
    pvalue: ESQLColumn;
  };
}

export interface ESQLAstCompletionCommand extends ESQLCommand<'completion'> {
  prompt: ESQLAstExpression;
  inferenceId: ESQLLiteral;
  targetField?: ESQLColumn;
}

export interface ESQLAstFuseCommand extends ESQLCommand<'fuse'> {
  fuseType?: ESQLIdentifier;
}

export interface ESQLAstRerankCommand extends ESQLCommand<'rerank'> {
  query: ESQLLiteral;
  fields: ESQLAstField[];
  targetField?: ESQLColumn;
  inferenceId: ESQLLiteral | undefined;
}

export interface ESQLAstForkCommand extends ESQLCommand<'fork'> {
  args: ESQLForkParens[];
}
// ...
export interface ESQLAstHeaderCommand<Name extends string = string, Arg = ESQLAstExpression>
  extends ESQLAstBaseItem {
  type: 'header-command';

  /** Name of the command */
  name: Name;

  /**
   * Represents the arguments for the command. It has to be a list, because
   * even the SET command was initially designed to accept multiple
   * assignments.
   *
   * Example:
   *
   * ```
   * SET setting1 = "value1", setting2 = "value2"
   * ```
   */
  args: Arg[];
}

export type ESQLAstSetHeaderCommand = ESQLAstHeaderCommand<
  'set',
  ESQLBinaryExpression<BinaryExpressionAssignmentOperator>
>;

export type ESQLIdentifierOrParam = ESQLIdentifier | ESQLParamLiteral;

export interface ESQLCommandOption extends ESQLAstBaseItem {
  type: 'option';
  args: ESQLAstItem[];
}
// ...
export interface ESQLSource extends ESQLAstBaseItem {
  type: 'source';
  sourceType: 'index' | 'policy';

  /**
   * Represents the prefix part of the source identifier. Empty string if not
   * present. Used in index pattern as the cluster identifier or as "mode" in
   * enrich policy.
   *
   * ```
   * FROM [<prefix>:]<index>
   * ```
   */
  prefix?: ESQLStringLiteral | undefined;

  /**
   * Represents the index part of the source identifier. Unescaped and unquoted.
   *
   * ```
   * FROM [<cluster>:]<index>
   * ```
   */
  index?: ESQLStringLiteral | undefined;

  /**
   * Represents the selector (component) part of the source identifier.
   *
   * ```
   * FROM <index>[::<selector>]
   * ```
   */
  selector?: ESQLStringLiteral | undefined;
}
// ...
export interface ESQLStringLiteral extends ESQLAstBaseItem {
  type: 'literal';

  literalType: 'keyword';

  value: string;
  valueUnquoted: string;

  /**
   * Whether the string was parsed as "unqouted" and/or can be pretty-printed
   * unquoted, i.e. in the source text it did not have any quotes (not single ",
   * not triple """) quotes. This happens in FROM command source parsing, the
   * cluster and selector can be unquoted strings:
   *
   * ```
   * FROM <cluster>:index:<selector>
   * ```
   */
  unquoted?: boolean;
}
// ...
export interface ESQLMessage {
  type: 'error' | 'warning';
  text: string;
  location: ESQLLocation;
  code: string;
  errorType?: 'semantic';
  requiresCallback?: 'getColumnsFor' | 'getSources' | 'getPolicies' | 'getJoinIndices' | string;
  underlinedWarning?: boolean;
}